Deprecate `brew cask update`.
